move email daemon code into its own file - NO OTHER CODE CHANGES, JUST MOVING STUFF
